+++ datatypes/src/jvmTest/kotlin/PublicKeyTest.kt
@@ -41,9 +41,9 @@ class PublicKeyTest : FreeSpec({
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
                 println(own.encodeToDer().encodeToString(Base16()))
-                println(own.multiBaseEncoded)
+                println(own.didEncoded)
                 own.encodeToDer() shouldBe pubKey.encoded
-                CryptoPublicKey.fromKeyId(own.multiBaseEncoded) shouldBe own
+                CryptoPublicKey.fromDid(own.didEncoded) shouldBe own
                 own.getJcaPublicKey().getOrThrow().encoded shouldBe pubKey.encoded
                 CryptoPublicKey.decodeFromTlv(Asn1Element.parse(own.encodeToDer()) as Asn1Sequence) shouldBe own
             }
@@ -89,7 +89,7 @@ class PublicKeyTest : FreeSpec({
 
                 println(Json.encodeToString(own))
                 println(own.iosEncoded.encodeToString(Base16()))
-                println(own.multiBaseEncoded)
+                println(own.didEncoded)
                 val keyBytes = ((ASN1InputStream(pubKey.encoded).readObject()
                     .toASN1Primitive() as ASN1Sequence).elementAt(1) as DERBitString).bytes
                 own.iosEncoded shouldBe keyBytes //PKCS#1